requirements.txt
astroid==2.4.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
dill==0.3.2
Django==2.2.2
django-cors-headers==3.4.0
django-rest-params==1.0.2
djangorestframework==3.11.0
imageio==2.8.0
isort==4.3.21
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lime==0.2.0.1
matplotlib==3.2.2
mccabe==0.6.1
networkx==2.4
numpy==1.19.0
Pillow==7.2.0
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
tqdm==4.47.0
typed-ast==1.4.1
wrapt==1.12.1
xgboost==1.1.1